extract logic into new DateCardSidebar component
